refactor: Score 객체 삭제
