--- src/main/java/domain/Score.java
@@ -1,19 +1,0 @@
-package domain;
-
-public final class Score {
-    private int choScore;
-    private int hanScore;
-
-    public Score(int choScore, int hanScore) {
-        this.choScore = choScore;
-        this.hanScore = hanScore;
-    }
-
-    public int getChoScore() {
-        return choScore;
-    }
-
-    public int getHanScore() {
-        return hanScore;
-    }
-}
